mms/requirements.txt
Django==1.8
django-filter==0.15.3
djangorestframework==3.5.3
factory-boy==2.7.0
fake-factory==0.7.2
geojson==1.3.3
ipaddress==1.0.17
Markdown==2.6.7
psycopg2==2.6.2
python-dateutil==2.6.0
six==1.10.0
